Kaggle_Store_Sales_SARIMAX_LSTM
In this notebook, we will explore the Corporación Favorita Grocery Sales Forecasting competition on Kaggle. The goal of this competition is to predict the unit sales for thousands of items sold at different Favorita stores located in Ecuador. We will use the data provided to train a model that can accurately forecast the sales for each item in each store.

To achieve our goals, we use trend features, moving averages, and SARIMAX to impute missing data. For the actual model training, we use  LightGBM to predict the sales. We also optimized our model using Optuna to find the best hyperparameters.